Rename module to remove unneeded underscore and run clang format
