+++ lib/index.js
@@ -1,3 +1,3 @@
 "use strict";
 
-module.exports = require('./vt_shaver.js');
+module.exports = require('./vtshaver.js');
