+++ src/vtshaver.cpp
